remove: 회원가입 컴포넌트 중 불필요한 파일 삭제
